requirements.txt
nltk==3.8.1
python-dotenv==1.0.0
requests==2.31.0 
